requirements.txt
Flask==1.1.2
gunicorn==20.0.4
numpy==1.19.1
scikit-learn==0.23.2
pandas==1.1.0
psycopg2-binary==2.8.5
xgboost==1.1.1